Utility: Collect and summarize metadata across subprojects
Adds helper script to collect metadata, process it, print a summary of
results, and save to disk.

Provides infrastructure for populating a pandas DataFrame with project-wide
metadata that we can expand on in the future for specific reporting and
evaluation needs.

Motivated by efforts to better understand HoloHub growth and maintenance.

Signed-off-by: Tom Birdsong <tbirdsong@nvidia.com>