Op: Add STL-to-USD Holoscan operator
Adds an operator to convert an STL-formatted mesh to UsdGeom format in
an active OpenUSD scene as part of a Holoscan workflow.

OpenUSD is an expansive format and ecosystem for scene description.
Connecting Holoscan to USD may facilitate future operations for
interoperability with the NVIDIA Omniverse cloud platform and other
applications.

Subsequent work will add a tutorial demonstrating how to use the
operator.

co-authored-by: Cristiana Dinea <cdinea@nvidia.com>
co-authored-by: Wendell Hom <whom@nvidia.com>
Signed-off-by: Tom Birdsong <tbirdsong@nvidia.com>