Build: Update benchmarking dependencies in container Dockerfile
Updates the default HoloHub container Dockerfile to install a complete
list of dependencies to use flow benchmarking tools.

The Holoscan Flow Benchmarking tutorial was updated in 86e9800 to
include the `pydot` and `xdot` Python packages in its requirements.
`pydot` has a transitive dependency on `graphviz`, while `xdot` has a
GUI dependency on GTK and canberra-gtk along with `graphviz`.

With these changes applied I was able to follow the benchmarking
tutorial to view visual graphs of my application flow latency via the
HoloHub container.

Signed-off-by: Tom Birdsong <tbirdsong@nvidia.com>