Infra: Detect custom Dockerfile in `build_and_run` command
Updates the `./dev_container build_and_run` command to detect and use a
custom Dockerfile provided by the application if one is available.

The motivation for this change is to increase the projects that can be
covered by the `build_and_run` command with minimal extra arguments on
the command line. The eventual goal of this work is to provide
build-and-run coverage for a maximum subset of HoloHub projects via the
HoloHub landing page.

Under previous behavior `build_and_run` supported using a custom docker
file only if explicitly passed as a command line argument and did not
enforce a discovery convention.

Under updated behavior `build_and_run` defines the following heuristics
to find a custom Dockerfile for a HoloHub application, in order of
priority:
1. Use a Dockerfile provided on the command line;
2. Use a Dockerfile provided in `metadata.json`;
3. Use a Dockerfile in the same folder as `metadata.json`;
4. If `metadata.json` is in a language subfolder, use a Dockerfile in
   the one-level-up project directory;
5. If none of the conditions above are satisfied, use the default
   HoloHub Dockerfile.

Signed-off-by: Tom Birdsong <tbirdsong@nvidia.com>