Infra: Update `gather_metadata` app name extraction procedure
Aligns the `gather_metadata.py` application name extraction procedure to
that used in the `run` script.

Under previous behavior `gather_metadata.py` assumed that the
folder directly below "applications" should be taken as the name of the
application. That behavior does not reflect standing HoloHub conventions
in file organization and in the `run` script where applications may be
nested, such as the `orsi` application containing several project
subfolders.

Updated behavior follows a similar procedure to the `run` script where
the subfolder containing `metadata.json` is usually taken as the
application name, or its parent is taken as such if `metadata.json` is
in a language subfolder.

Signed-off-by: Tom Birdsong <tbirdsong@nvidia.com>
Signed-off-by: Mikael Brudfors <mbrudfors@nvidia.com>