Infra: Workaround for `docker build` failure for older Holoscan SDK
Implements a workaround to address an observed `docker build` failure
for the Holoscan SDK v0.6.0 base container image.

@grelee reported an issue where the `dev_container` command would fail
to build if the 0.6.0 container image was used:

```sh
./dev_container build --base_img nvcr.io/nvidia/clara-holoscan/holoscan:v0.6.0-dgpu
```

That command would fail to install `xdot` and indicated that
`setuptools` was not found. As a workaround we can install `setuptools`
in advance conditionally if we are building on an older Ubuntu version.

The resulting `xdot` appears to not be usable and more work is required
to fully address the issue. But, this workaround allows the container
build to succeed so that the user can continue with HoloHub usage as
normal.

Signed-off-by: Tom Birdsong <tbirdsong@nvidia.com>
Signed-off-by: Mikael Brudfors <mbrudfors@nvidia.com>