Support data export in out of body detection app
This patch modifies out of body detection application to
support exporting application output to the CSV files.
This output can later be used by analytics applications.

This patch adds a new flag `enable_analytics` in config file
which can be used to toggle exporting output to the CSV file.
It is by default set to false.

Signed-off-by: Arun Patole <apatole@nvidia.com>