Infra: Add `benchmarks` folder support in `run` commands
Adds infrastructure support to list, build, and run projects in the
`benchmarks` folder with the `run` script.

f458d3bedbd598eba06887049421453a54501458 created the "benchmarks" folder
and project schema to highlight HoloHub performance-focused tools and
projects. However, the "run" script at that time was hardcoded to search
for projects only in the "applications" folder.

This change fixes the `model_benchmarking` application in particular to
run according to existing README instructions.

Signed-off-by: Tom Birdsong <tbirdsong@nvidia.com>

Signed-off-by: Rony Rado <rrado@nvidia.com>