App: Revert `volume_rendering` Holoscan SDK minimum to v1.0
Reverts `volume_rendering` C++/Python application and `volume_loader`
operator to require a minimum of Holoscan SDK v1.0, rather than v2.1.

As part of Volume Rendering Python updates
663542b67da39a32a2ee0d37d9b91aefb0e787e7 increased the minimum version
of Holoscan SDK required for the apps and operator to v2.1. However,
`volume_rendering` C++ is required to be compatible with Holoscan SDK v1.0
and v2.0 as part of the Nvidia AI Enterprise program strategy documented
at
https://github.com/nvidia-holoscan/holohub/blob/main/doc/release.md#holoscan-sdk-long-term-compatibility.

On review the C++ `volume_rendering` application still builds and runs
with Holoscan SDK v1.0 and v2.0 NVAIE containers if the CMake minimum
package requirement is reverted to v1.0. Proposing we revert that
minimum requirement with no other feature changes.

`volume_rendering` Python is observed to be compatible with v2.0+ and is
amended as such.

Signed-off-by: Tom Birdsong <tbirdsong@nvidia.com>