Benchmark: Add Release Benchmarking tutorial
Adds a benchmarking tutorial workflow to collect and present
benchmarking statistics for a Holoscan SDK release.

The tutorial aims to define a reproducible workflow based on existing
Holoscan SDK and HoloHub tools that can be run with each Holoscan SDK
release. Includes container with reporting tools and instructions for
how to reproduce each component of the workflow.

Signed-off-by: Tom Birdsong <tbirdsong@nvidia.com>