Add psd_pipeline_sim application
When trying to use Holoscan as an SDR framework (like GNU Radio or
REDHAWK), there are a few missing pieces such as a generic IQ file
reader application and simple blocks like FFT and PSD. It also lacks
some support for VITA 49.

Add an application which takes in file data, does some signal
processing, and outputs VITA 49 spectral data packets.

In the future, we plan on adding a VITA 49 ingest operator which
leverages ANO, but for now, the file-based simulation is helpful.

Signed-off-by: John Moon <john.moon@vts-i.com>