Enforce use of jammy for github actions
Need to install some deps with `apt install` and pinned
versions, so pinning the ubuntu release version is needed.

Addresses:
```
[command] apt install --no-install-recommends -y clang-format=1:14.0*

WARNING: apt does not have a stable CLI interface. Use with caution in scripts.

Reading package lists...
Building dependency tree...
Reading state information...
Package clang-format is not available, but is referred to by another package.
This may mean that the package is missing, has been obsoleted, or
is only available from another source

E: Version '1:14.0*' for 'clang-format' was not found
```

Signed-off-by: Alexis Girault <agirault@nvidia.com>