Build: Set default `CUDA_ARCHITECTURES` for builds
Set a default `CMAKE_CUDA_ARCHITECTURES` list for builds to inherit that covers
most modern architectures.

The list here is at parity with Holoscan SDK build defaults. Subprojects
in HoloHub may choose to override this value with their own default settings
or set `CUDA_ARCHITECTURES` directly on targets.

Resolves an observed issue on IGX 1.1 where not setting this property
resulted in a default value of "52" in the Multi-AI Ultrasound application,
leading to a "PTX" failure message. With this fix the Multi-AI Ultrasound
application inherits the updated list of compute capabilities by default.

Note that existing builds may need to clear the `CMakeCache.txt` or the
specific architectures cache entry to apply the fix.

co-authored-by: Wendell Hom <whom@nvidia.com>
Signed-off-by: Tom Birdsong <tbirdsong@nvidia.com>