Switch to container-based install of MatX
Currently, all of the operators in the PSD pipeline pull in their own
version of MatX. In some scenarios, we'd like to patch the version of
MatX we're using on all of these operators.

Instead of updating the CMake logic to apply patches in several
different places, let's just move to installing MatX in the development
container and apply out patches then. That way, we know each operator in
the pipeline is using the same version of MatX.

The patches applied in this commit:

- Fix a memory leak (applied to upstream MatX)
- Update the version of dlpack.h to match Holoscan's bundled version

Signed-off-by: John Moon <john.moon@vts-i.com>