Update psd_pipeline to align with new ANO API
A recent commit (b8b7c82) removed the ANO operator interface which the
psd_pipeline application uses.

Instead, the ANO is now exposed as a library which we initialize and
fetch bursts from.

Before this change, the VITA connector compute() function was only called
when individual batches were ready in the ANO. Now, the function is run
in a loop and can process any/all channels it wants.

This change requires a refactor in how the psd_pipeline's VITA connector
operator behaves; now, it kicks off copy/processing for all channels it
can and emits the channels one at a time (as one compute() call cannot
have multiple op_output.emit() calls).

Signed-off-by: John Moon <john.moon@vts-i.com>