Reduce psd_pipeline log spew
Currently, the PSD pipeline prints a lot of logs. It prints when
indivdual channel packet batches are filled, when PSD packets are sent
out the end, and every context packet's contents.

Instead, let's reduce the log spew by moving most of those prints to
`HOLOSCAN_LOG_DEBUG`.

Still, we want to see some status, so update the packetizer operator to
print a measure of how long it takes to send N (configurable) packets
and only print context information if the context changes.

Signed-off-by: John Moon <john.moon@vts-i.com>