Isaac Sim Holoscan bridge
The Isaac Holoscan Bridge is a sophisticated application that demonstrates
the integration between NVIDIA's Isaac Simulator and the Holoscan framework.

It creates a bridge where simulated data, including camera and robotic
joint data, from Isaac Sim is processed through a Holoscan pipeline.
The application features a virtual camera in Isaac Sim that captures video
frames, which are then streamed to a Holoscan pipeline for real-time processing.
The processed data is then again displayed in Isaac Sim. The application also
streams joint positions from a robot to a Holoscan pipeline. The pipeline
modfies thes joint positions and feeds them back into the simulation moving
the robotic arm.

Signed-off-by: Andreas Heumann <aheumann@nvidia.com>