App: Remove extraneous "colonoscopy_segmentation" test PYTHONPATH
Followup to e86a13e removing GXF_LIB_DIR from test PYTHONPATH. As
of f9d44deb the Holoscan Python package is installed in the Python
dist-packages path at container build time, so we do not need to set
PYTHONPATH to find it.

The app with replayer preset builds no Python operators, so the
HoloHub Python module is not required.

Signed-off-by: Tom Birdsong <tbirdsong@nvidia.com>