Ext: Disable FP16 TensorRtInference generation in latest TensorRT (#1120)
Explicitly disable FP16 support in the GXF TensorRtInference extension
when building with the latest TensorRT 10.13 distribution.

FP16 is deprecated and scheduled for removal.

Addresses an observed issue where building the extension with FP16
support results in unreliable segmentation output in the Endoscopy Tool
Tracking application, leading to a missing overlay when the application
runs on Jetson AGX Thor.

Signed-off-by: Tom Birdsong <tbirdsong@nvidia.com>