Disable `HOLOHUB_BUILD_PYTHON` for C++ (#1201)
By default, `HOLOHUB_BUILD_PYTHON` is `ON`, even when using `--language
cpp`. This will build python bindings which are not needed when building
a C++ application or operator. This PR turn it `OFF` if the language is
not Python.

<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

## Bug Fixes
* Improved build handling for non-Python projects by ensuring Python
build components are properly disabled when building projects in other
languages.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->

Signed-off-by: B Hashemian <3968947+bhashemian@users.noreply.github.com>
Signed-off-by: eagostini <eagostini@nvidia.com>