avoid possible segfault on shutdown of the Python version of the `cvcuda_basic` application (#1214)
A segfault related to an `nvidia::gxf::MemoryBuffer` object is currently
observed during application shutdown when a user closes the display
window of the `cvcuda_basic` Python application. I suspect both CV-CUDA
and Holoscan are trying to free the same memory.

Making an explicit copy of the input tensor via CuPy resolves the
segfault. We will need to look into the CV-CUDA code in more detail to
see if we can avoid the need for this copy, but in the meantime it seems
best to merge this change.

<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

## Bug Fixes
- Fixed an application crash that occurred during shutdown caused by
conflicting memory operations in CV-CUDA tensor processing.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->

Signed-off-by: Gregory R. Lee <grelee@nvidia.com>
Co-authored-by: Bruce Hashemian <3968947+bhashemian@users.noreply.github.com>
Signed-off-by: eagostini <eagostini@nvidia.com>