Remove lint.sh per PR feedback
linting is not a good example of CLI dev script enhancements. We already
have a path for linting with `./holohub lint --install-dependencies` and
`./holohub lint [--fix]`, and a linting container may not need HoloHub
dev dependencies as available in the HoloHub default Dockerfile.

We could revisit a containerized linting approach in the future with a
minimal Dockerfile and custom handling, but it is out of scope and low
priority right now.

Signed-off-by: Tom Birdsong <tbirdsong@nvidia.com>